Different LED modes as information for drivers
